+++ AutoGate/docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # Release notes
 ## Upcoming release
-* Updated for SMAPI 3.0 and Stardew Valley 1.3.33.
+* Updated for Stardew Valley 1.5 (including split-screen mode).
 * Gates now open immediately when approached, without a half-second delay.
 * Added update keys.
 * Internal refactoring.